fixes for collection signals; align_type wip

+++ formative/signals.py
@@ -153,32 +153,39 @@ def formblock_post_save(sender, instance, created, raw, **kwargs):
     if not new and block._old_name != block.name:
         delete_sub_labels(block.form, block._old_name)
 
+@receiver(pre_save, sender=CollectionBlock)
+def customblock_pre_save(sender, instance, raw, **kwargs):
+    if raw: return
+    
+    if instance.pk:
+        orig = CollectionBlock.objects.get(pk=instance.pk)
+        instance._old_name = orig.name
+
 @receiver(post_save, sender=CollectionBlock)
 def collectionblock_post_save(sender, instance, created, raw, **kwargs):
-    block = instance
+    block, new = instance, created
     if raw: return
     
     fields = block.collection_fields()
     
-    if created:
-        existing = block.form.custom_blocks().filter(name__in=fields, page=0)
-        existing_names = existing.values_list('name', flat=True)
-        for name in fields:
-            if name in existing_names: continue
-        
-            item_field = CustomBlock.text_create(form=block.form, name=name,
-                                                 page=0)
-            item_field.save()
+    existing = block.form.custom_blocks().filter(name__in=fields, page=0)
+    existing_names = existing.values_list('name', flat=True)
+    for name in fields:
+        if name in existing_names: continue
+    
+        item_field = CustomBlock.text_create(form=block.form, name=name, page=0)
+        item_field.save()
     
     refs = CollectionBlock.objects.filter(any_name_field(_=OuterRef('name')),
                                           form=block.form)
-    block.form.collections().filter(~Exists(refs), page=0, _rank__gt=1).delete()
-    
-    if not created: return
+    block.form.custom_blocks().filter(~Exists(refs),
+                                      page=0, _rank__gt=1).delete()
     
     text, style = default_text(block.name) + ':', FormLabel.LabelStyle.VERTICAL
+    paths = []
     l = FormLabel.objects.get_or_create(form=block.form, path=block.name,
                                         defaults={'style': style, 'text': text})
+    paths.append(l[0].path)
     
     if block.fixed: return # no implementation yet for fixed + collection
     style = FormLabel.LabelStyle.WIDGET
@@ -187,6 +194,11 @@ def collectionblock_post_save(sender, instance, created, raw, **kwargs):
         l = FormLabel.objects.get_or_create(form=block.form, path=path,
                                             defaults={'style': style,
                                                       'text': text})
+        paths.append(l[0].path)
+    
+    delete_sub_labels(block.form, block.name, paths)
+    if not new and block._old_name != block.name:
+        delete_sub_labels(block.form, block._old_name)
 
 def delete_sub_labels(form, name, exclude=[]):
     sl = form.labels.filter(path__startswith=name+'.').exclude(path__in=exclude)
@@ -213,13 +225,17 @@ def collectionblock_post_delete(sender, instance, **kwargs):
                                path__endswith='_')).delete()
     
     blocks = block.form.collections(name=block.name)
-    if not blocks:
-        delete_block_labels(block.form, block.name)
-        return
+    if not blocks: delete_block_labels(block.form, block.name)
+    else:
+        for name in block.collection_fields():
+            if not blocks.filter(any_name_field(_=name)):
+                block.form.labels.filter(path='.'.join((block.name,
+                                                        name))).delete()
     
-    for name in block.collection_fields():
-        if not blocks.filter(any_name_field(_=name)):
-            block.form.labels.filter(path='.'.join((block.name, name))).delete()
+    refs = CollectionBlock.objects.filter(any_name_field(_=OuterRef('name')),
+                                          form=block.form)
+    block.form.custom_blocks().filter(~Exists(refs),
+                                      page=0, _rank__gt=1).delete()
 
 
 app_cache = {}
